Specify auth parameters in devstack tests
Single-node devstack does not have SSH key for stack user.
In all tests we will use custom os-faults key and connect to localhost
as `stack` user. Tests are also executed as `stack` user explicitly
(by default they are run as zuul user, but we do not want to rely on this).

Change-Id: Ibcc9a406dc850363e651d4023874147db45214f7